+++ Public/Invoke-DotCoverForExecutable.ps1
@@ -17,6 +17,8 @@
   Attribute coverage filters for dotCover, to indicate what should and should not be covered.
 .PARAMETER ProcessFilters
   Process coverage filters for dotCover, to indicate what should and should not be covered. Requires Dot Cover 2016.2 or later.
+.PARAMETER TargetWorkingDirectory
+  The working directory of the target executable.
 #>
 function Invoke-DotCoverForExecutable {
   [CmdletBinding()]
@@ -41,7 +43,10 @@ function Invoke-DotCoverForExecutable {
     [string] $AttributeFilters = '',
 
     [Parameter(Mandatory = $False)]
-    [string] $ProcessFilters = ''
+    [string] $ProcessFilters = '',
+
+    [Parameter(Mandatory = $False)]
+    [string] $TargetWorkingDirectory
   )
 
   $DotCoverArguments = @(
@@ -66,6 +71,10 @@ function Invoke-DotCoverForExecutable {
   if ($TargetArguments) {
     $EscapedTargetArguments = ConvertTo-ShellEscaped $TargetArguments
     $DotCoverArguments += "/TargetArguments=$EscapedTargetArguments"
+  }
+
+  if ($TargetWorkingDirectory) {
+    $DotCoverArguments += "/TargetWorkingDir=`"$TargetWorkingDirectory`""
   }
 
   $DotCoverPath = Get-DotCoverExePath -DotCoverVersion $DotCoverVersion